+++ run_tests.py
@@ -9,8 +9,8 @@ To run the tests:
     cd intelligent-music-recommender
     python ./run_tests.py
 """
-from tests.test_knowledge_base_api import TestMusicKnowledgeBaseAPI
-from tests.test_db_schema import TestDbSchema
+from tests.knowledge_base.test_knowledge_base_api import TestMusicKnowledgeBaseAPI
+from tests.knowledge_base.test_db_schema import TestDbSchema
 
 if __name__ == '__main__':
     unittest.main()
