+++ tests/knowledge_base/test_db_schema.py
